# internal/contrato/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "notarum",
    "version": "1.0.0",
    "description": "API abierta de sólo lectura del Boletín Oficial de la República Argentina. Entrega el calendario de ediciones, cada edición por sección y fecha con sus avisos, cada aviso con su texto, los anexos en PDF y el catálogo de rubros. No opina, no clasifica, no puntúa: entrega el dato crudo y bien tipado.\n\nSin clave para leer. Límite de pedidos por IP. Los datos son del Boletín Oficial; notarum es sólo una caché legible.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "notarum",
      "url": "https://github.com/diegoparras/notarum"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "esta instancia"
    }
  ],
...
  "paths": {
...
    "/v1/calendario/{anio}/{seccion}": {
      "get": {
        "tags": [
          "catalogo"
        ],
        "summary": "Días con edición de un año",
        "operationId": "verCalendario",
        "parameters": [
          {
            "name": "anio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1990
            },
            "example": 2026
          },
          {
            "$ref": "#/components/parameters/Seccion"
          }
        ],
        "responses": {
          "200": {
            "description": "El calendario del año",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendario"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
    "/v1/ediciones/{seccion}/{fecha}": {
      "get": {
        "tags": [
          "ediciones"
        ],
        "summary": "Edición de una sección en una fecha",
        "description": "El sumario completo de ese día. Si ese día no hubo edición devuelve 404 con `sin_edicion: true`, que no es una falla.",
        "operationId": "verEdicion",
        "parameters": [
          {
            "$ref": "#/components/parameters/Seccion"
          },
          {
            "$ref": "#/components/parameters/Fecha"
          },
          {
            "name": "rubro",
            "in": "query",
            "required": false,
            "description": "Filtra por nombre de rubro; acepta el nombre exacto o un prefijo, sin distinguir mayúsculas.",
            "schema": {
              "type": "string"
            },
            "example": "DECRETOS"
          }
        ],
        "responses": {
          "200": {
            "description": "La edición",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Edicion"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "404": {
            "description": "Ese día no hubo edición",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
    "/v1/ediciones/{seccion}": {
      "get": {
        "tags": [
          "ediciones"
        ],
        "summary": "Resúmenes de un rango de fechas",
        "description": "Devuelve un resumen por día (sin avisos) para las ediciones que ya están en la caché local. Las que todavía no se bajaron se listan en `faltantes`: se llenan con `notarum rellenar`, porque la API no baja un año entero adentro de un pedido HTTP. El tope del rango son 366 días.",
        "operationId": "verRango",
        "parameters": [
          {
            "$ref": "#/components/parameters/Seccion"
          },
          {
            "name": "desde",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-09-01"
          },
          {
            "name": "hasta",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-09-30"
          }
        ],
        "responses": {
          "200": {
            "description": "Los resúmenes disponibles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rango"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
    "/v1/avisos/{seccion}/{id}/{fecha}": {
      "get": {
        "tags": [
          "avisos"
        ],
        "summary": "Un aviso con su texto completo",
        "operationId": "verAviso",
        "parameters": [
          {
            "$ref": "#/components/parameters/Seccion"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "El id del aviso. En la primera sección es numérico; en la segunda y la tercera puede ser alfanumérico (por ejemplo A1522579).",
            "schema": {
              "type": "string"
            },
            "example": "346633"
          },
          {
            "$ref": "#/components/parameters/Fecha"
          }
        ],
        "responses": {
          "200": {
            "description": "El aviso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detalle"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "404": {
            "$ref": "#/components/responses/NoEncontrado"
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
    "/v1/anexos/{seccion}/{nro}/{id}/{fecha}": {
      "get": {
        "tags": [
          "avisos"
        ],
        "summary": "El PDF de un anexo",
        "description": "Devuelve el archivo tal como lo publica el Boletín. La ruta es la que viene en el campo `url` de cada anexo del detalle; admite el sufijo `.pdf` en la fecha.",
        "operationId": "verAnexo",
        "parameters": [
          {
            "$ref": "#/components/parameters/Seccion"
          },
          {
            "name": "nro",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "12"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "7756488"
          },
          {
            "name": "fecha",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "20260901.pdf"
          }
        ],
        "responses": {
          "200": {
            "description": "El PDF",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
    "/v1/rubros/{seccion}": {
      "get": {
        "tags": [
          "catalogo"
        ],
        "summary": "Catálogo de rubros de una sección",
        "operationId": "verRubros",
        "parameters": [
          {
            "$ref": "#/components/parameters/Seccion"
          }
        ],
        "responses": {
          "200": {
            "description": "Los rubros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rubro"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
    "/v1/buscar": {
      "get": {
        "tags": [
          "avisos"
        ],
        "summary": "Búsqueda por texto y fecha",
        "description": "Busca avisos por texto y fecha. Con el motor de almacenamiento sqlite hay un índice local: `fuente=indice` busca sobre él sin pedirle nada al Boletín (más rápido, sin tope de rango); `fuente=sitio` siempre consulta la búsqueda avanzada del Boletín; `fuente=auto` (por defecto) usa el índice cuando tiene historia del rango y si no va al sitio. La respuesta dice cuál se usó.",
        "operationId": "buscar",
        "parameters": [
          {
            "name": "seccion",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "primera",
                "segunda",
                "tercera"
              ]
            }
          },
          {
            "name": "texto",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "decreto"
          },
          {
            "name": "desde",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-09-01"
          },
          {
            "name": "hasta",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-09-03"
          },
          {
            "name": "rubro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "fuente",
            "in": "query",
            "required": false,
            "description": "De dónde traer los resultados.",
            "schema": {
              "type": "string",
              "enum": [
                "auto",
                "indice",
                "sitio"
              ],
              "default": "auto"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Resultados por página; sólo con fuente indice.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            }
          },
          {
            "name": "todas",
            "in": "query",
            "required": false,
            "description": "Exigir todas las palabras; sólo con fuente sitio.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Los resultados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Busqueda"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PedidoInvalido"
          },
          "502": {
            "$ref": "#/components/responses/FalloDelSitio"
          }
        }
      }
    },
...
    "/v1/provincial": {
      "get": {
        "tags": [
          "provincial"
        ],
        "summary": "Buscar normativa provincial",
        "description": "Busca en las 81 mil normas de las 24 provincias. Sin filtros devuelve todo el catálogo paginado, de la más nueva a la más vieja. Cuando hay texto, primero va lo que se llama así y después lo que lo menciona de paso.",
        "operationId": "buscarProvincial",
        "parameters": [
          {
            "name": "texto",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Palabras a buscar en el título y las materias. No hace falta poner los acentos; todas las palabras tienen que estar."
          },
          {
            "name": "provincia",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Nombre, código INDEC o prefijo. Sin esto busca en las 24."
          },
          {
            "name": "tipo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Ley, Decreto Ley, Constitución Provincial… La lista está en /v1/provincial/tipos."
          },
          {
            "name": "desde",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Año de sanción más viejo."
          },
          {
            "name": "hasta",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Año de sanción más nuevo."
          },
          {
            "name": "vigentes",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Con 1 deja fuera lo derogado, lo caduco y las modificatorias."
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limite",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Las normas que cumplen.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoProvincial"
                }
              }
            }
          },
          "400": {
            "description": "Una provincia que no existe, o un año mal escrito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Esta instancia todavía no bajó el catálogo provincial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/provincial/{id}": {
      "get": {
        "tags": [
          "provincial"
        ],
        "summary": "Una norma provincial",
        "description": "Trae una norma por su identificador de SAIJ.",
        "operationId": "verNormaProvincial",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identificador de SAIJ, como LPB1000000.",
            "example": "LPB1000000"
          }
        ],
        "responses": {
          "200": {
            "description": "La norma, con el enlace a su ficha en SAIJ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormaProvincial"
                }
              }
            }
          },
          "404": {
            "description": "No hay una norma con ese identificador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Esta instancia todavía no bajó el catálogo provincial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Seccion": {
        "name": "seccion",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "primera",
            "segunda",
            "tercera"
          ]
        },
        "example": "primera"
      },
      "Fecha": {
        "name": "fecha",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "date"
        },
        "example": "2026-09-01"
      }
    },
    "responses": {
      "PedidoInvalido": {
        "description": "El pedido está mal armado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NoEncontrado": {
        "description": "No existe",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "FalloDelSitio": {
        "description": "El Boletín Oficial no contestó o cambió de forma",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Calendario": {
        "type": "object",
        "required": [
          "anio",
          "seccion",
          "fechas"
        ],
        "properties": {
          "anio": {
            "type": "integer"
          },
          "seccion": {
            "type": "string"
          },
          "fechas": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          },
          "con_suplemento": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          }
        }
      },
      "Rubro": {
        "type": "object",
        "required": [
          "id",
          "nombre"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
...
    }
  }
}